+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/Interfaces/ISemanticTokenEditHandler.cs
@@ -4,11 +4,12 @@
 using System;
 using MediatR;
 using Microsoft.AspNetCore.Razor.LanguageServer.Common;
+using Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Models;
 using OmniSharp.Extensions.JsonRpc;
 using OmniSharp.Extensions.LanguageServer.Protocol;
 using OmniSharp.Extensions.LanguageServer.Protocol.Client.Capabilities;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Interfaces
 {
     public class SemanticTokensEditParams : IRequest<SemanticTokensOrSemanticTokensEdits?>
     {
