+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/Interfaces/ISemanticTokenLegendHandler.cs
@@ -3,10 +3,11 @@
 
 using MediatR;
 using Microsoft.AspNetCore.Razor.LanguageServer.Common;
+using Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Models;
 using OmniSharp.Extensions.JsonRpc;
 using OmniSharp.Extensions.LanguageServer.Protocol;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Interfaces
 {
     [Method(LanguageServerConstants.RazorSemanticTokenLegendEndpoint)]
     internal interface ISemanticTokenLegendHandler :
