+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/Models/SemanticTokenLegendParams.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic
+using MediatR;
+
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Models
 {
-    public class SemanticTokens
+    public class SemanticTokenLegendParams : IRequest<SemanticTokenLegendResponse>, IBaseRequest
     {
-        public string ResultId { get; set; }
-        public uint[] Data { get; set; }
     }
 }
