+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/Models/SemanticTokenParams.cs
@@ -3,14 +3,12 @@
 
 using System;
 using MediatR;
-using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Models
 {
-    public class SemanticTokensRangeParams : IRequest<SemanticTokens>
+    public class SemanticTokenParams : IRequest<SemanticTokens>
     {
         public RazorLanguageKind Kind { get; set; }
         public Uri RazorDocumentUri { get; set; }
-        public Range Range { get; set; }
     }
 }
